+++ src/tar_format_types.rs
@@ -145,7 +145,7 @@ impl<const N: usize, const R: u32> TarFormatNumber<N, R> {
     where
         T: num_traits::Num,
     {
-        let str = self.0.as_str_until_first_space().unwrap_or("0");
+        let str = self.0.as_str_until_first_space().unwrap_or("");
         T::from_str_radix(str, R)
     }
 
@@ -160,12 +160,9 @@ impl<const N: usize, const R: u32> Debug for TarFormatNumber<N, R> {
     fn fmt(&self, f: &mut Formatter) -> core::fmt::Result {
         let sub_array = &self.0.bytes[0..self.0.size()];
         match self.as_number::<u64>() {
-            Err(msg) => write!(f, "{}", msg),
-            Ok(val) => write!(f, "{}", val),
+            Err(msg) => write!(f, "{} [{:?}]", msg, from_utf8(sub_array)),
+            Ok(val) => write!(f, "{} [{:?}]", val, from_utf8(sub_array)),
         }?;
-        if let Ok(sub_array) = from_utf8(sub_array) {
-            write!(f, " [{}]", sub_array)?;
-        }
         write!(f, " [{:?}]", self.0.bytes)
     }
 }
@@ -329,5 +326,13 @@ mod tar_format_number_tests {
         let str = [b'0', b'1', b'0', b' ', 0];
         let str = TarFormatNumber::<5, 10>::new(str);
         assert_eq!(str.as_number::<u64>(), Ok(10));
+    }
+
+    #[test]
+    fn test_as_number_with_invalid_utf8() {
+        let str = TarFormatNumber::<2, 10>::new([0xff, 0]);
+        assert!(str.as_number::<u64>().is_err());
+
+        let _ = format!("{str:?}");
     }
 }